rename Query to QueryEditor
